add debugging, finalize get_command method
Add debugging prints outs to each method. Debugging defaults are set to
False. To use debugging, set debug=True on the method that you call.

Finalized the get_command method so that you can use it generically for
any call that requires HTTP GET.

Add visual studio code launch.json so that efficient launches of python
can be used (so it does nto pause on entry.)