CLI Args
Added teh ability to use command line arguments. Also rebuilt the debug
system so that if the debug argument is specified it will properly
output the debug information in a safe way. Added a shebang so that it
can be run as a standalone program on a *nix based system. Added a
docstring to the file itself for complience purposes.