Add More Caching
What is better than fast? Even faster, that is what. With this update
even more caching has been added. This time to the third authentication
step. REST access is now cached for the specified ammount when first
loggin into the API. The .gitignore file has been updated with this in
mind.